2024/06/03
I added multi threading to MACOS but it is still incredibly slow! And I figured out why, it's because Core Graphics itself. Now I am gonna add OpenGL support to MACOS version, if it is not supporting OpenGL (Below MacOS 10.14.X versions), then it will try to compile the project with Core Graphics. I know it is silow but it's honestly works lol. TODO: Add OpenGL