2025/04/04
Umm... Well... Idk where to start. In Windows, fixed a bug about the mouse cross is not turning default when the mouse is hovering on top of the window. Renamed all of the functions and the library itself because just 'GRAPHICS' is not a cool name. Moved the submodule LIBCMD. Soon, I'll add a lot of functions and fix the segmentation fault issue that's happening in X11. And will fix some setting issues that's not happening on Windows' window management. Also for test my grpahic library, I created a simply 3D block rendering stuff.